requirements.txt
# # Core dependencies
# # python==3.11.0                    # Impossible à installer

# # MySQL CDC
# mysql-connector-python==8.2.0       # Connexion MySQL
# pymysql-replication==0.45           # Lecture du binlog (CDC)

# # Snowflake
# snowflake-connector-python==3.11.0   # Connexion Snowflake # lib instalée par setup.sh en amont (ce n'est plus d'actualité)
# snowflake-connector-python[pandas]==3.7.2
# snowflake-sqlalchemy==1.5.1

# # Kafka
# kafka-python==2.0.2

# # Data processing
# pandas==2.1.4
# pyarrow==14.0.1
# polars==0.20.3

# # Configuration
# pyyaml==6.0.1                       # Lecture fichiers config
# python-dotenv==1.0.0

# # PII masking
# cryptography==41.0.7
# hashlib

# # Logging & monitoring
# python-json-logger==2.0.7
# colorlog==6.8.0

# # Testing
# pytest==7.4.3
# pytest-cov==4.1.0
# pytest-mock==3.12.0

# # Code quality
# black==23.12.1
# pylint==3.0.3
# flake8==6.1.0
# mypy==1.7.1

# # dbt (Transformation)
# dbt-core==1.8.0
# dbt-snowflake==1.8.0

# # Orchestration (optionnel)
# apache-airflow==2.7.3

# # Utilities
# requests==2.31.0
# click==8.1.7
# tqdm==4.66.1

# -------------------- Corrigé

# Core dependencies
dbt-core==1.8.0
dbt-snowflake==1.8.0
snowflake-connector-python[pandas]==3.12.4  # write_pandas() activé

# MySQL RDS
mysql-connector-python==8.4.0

# Kafka
kafka-python==2.0.2

# Data processing
pandas==2.1.4
pyarrow==14.0.1

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Utilities
requests==2.31.0
